# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Guzo Orthodox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin+)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUpdateUserBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /admin/audit:
    get:
      operationId: adminListAudit
      tags: [admin]
      summary: List audit log entries (admin+)
      parameters:
        - { name: actorId, in: query, schema: { type: integer } }
        - { name: action, in: query, schema: { type: string } }
        - { name: targetType, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditLogEntry" }
...
  /admin/system/settings:
    get:
      operationId: adminGetSystemSettings
      tags: [admin]
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SystemSetting" }
    patch:
      operationId: adminUpdateSystemSetting
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSystemSettingBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemSetting" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [user, moderator, admin, superadmin] }
        status: { type: string, enum: [active, suspended, banned] }
        suspendedUntil:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, email, name, role, status, notes, createdAt]
...
    AdminUpdateUserBody:
      type: object
      properties:
        name: { type: string }
        role: { type: string, enum: [user, moderator, admin, superadmin] }
        status: { type: string, enum: [active, suspended, banned] }
        suspendedUntil:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        notes: { type: string }
    AuditLogEntry:
      type: object
      properties:
        id: { type: string }
        actorId:
          oneOf:
            - { type: integer }
            - { type: "null" }
        actorEmail: { type: string }
        actorRole: { type: string }
        action: { type: string }
        targetType: { type: string }
        targetId: { type: string }
        metadata: { type: object, additionalProperties: true }
        ip: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, actorEmail, actorRole, action, targetType, targetId, metadata, ip, createdAt]
...
    SystemSetting:
      type: object
      properties:
        key: { type: string }
        value: {}
        description: { type: string }
        updatedAt: { type: string, format: date-time }
        updatedBy:
          oneOf:
            - { type: integer }
            - { type: "null" }
      required: [key, value, description, updatedAt]
    UpdateSystemSettingBody:
      type: object
      properties:
        key: { type: string }
        value: {}
        description: { type: string }
      required: [key, value]
...